separate rehype plugin for nuxt
